pyproject.toml
[project]
name = "fashion"
version = "0.1.0"
description = "H&M Personalized Fashion Recommendations"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dvc[s3]>=3.64.2",
    "fire>=0.7.1",
    "hydra-core>=1.3.2",
    "lightning>=2.6.0",
    "torch>=2.9.1",
    "torchvision>=0.22.0",
    "wandb>=0.23.1",
    "rich>=14.0.0",
    "pandas>=2.2.0",
    "numpy>=2.0.0",
    "scikit-learn>=1.6.0",
    "omegaconf>=2.3.0",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
]

[project.scripts]
fashionctl = "fashion.cli:main"

[tool.uv]
package = true
index-url = "https://pypi.org/simple"

[dependency-groups]
dev = [
    "kaggle>=1.8.2",
    "ruff>=0.14.8",
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["fashion"]
